+++ client/src/components/PlaylistCard.tsx
@@ -1,4 +1,5 @@
-import React, { useState, useEffect } from "react";
+import React, { useState, useEffect, useMemo } from "react";
+import axios from "axios";
 import { makeStyles, Theme, createStyles } from "@material-ui/core/styles";
 import Modal from "@material-ui/core/Modal";
 import Card from "@material-ui/core/Card";
@@ -13,20 +14,93 @@ import { useAppDispatch, useAppSelector } from "../redux/hooks";
 //  import reducer and state selector
 import { selectMusicApp, setPlaylists } from "../redux/musicAppSlice";
 
-import {PlaylistModal} from './PlaylistModal';
-// interface PlaylistProps{
-//     playlist: Playlist;
-// }
 
 const PlaylistCard = ({ playlistName }: any, { playlistId }: any) => {
-  
+  const musicAppState = useAppSelector(selectMusicApp);
+  const dispatch = useAppDispatch();
   const inputRef = React.createRef();
+  const [modalStyle] = React.useState(getModalStyle);
+ 
 
+  const getPlaylistSongs = () => {
+    const queryString = `http://localhost:8080/api/read/playlist/song`;
+
+    const body = {
+      params: {
+        playlistId: musicAppState.playlist.playlistId,
+      },
+    };
+    axios
+      .get(queryString, body)
+      .then((response) => {
+        console.log("response", response);
+        // const playlistData = response.data;
+        //  Dispatch the setPlaylistData reducer to save data to the state
+        // dispatch(setPlaylists(playlistData));
+        //  Check your redux devtools to see the data
+      })
+      .catch((error) => {
+        console.log("There was an error: ", error);
+      });
+  };
+
+  useMemo(() => {
+    getPlaylistSongs();
+  }, []);
+  
   const handleClick = () => {
     alert("You clicked on " + playlistName + " playlist.");
   };
+  const [open, setOpen] = React.useState(false);
 
+  const handleOpen = () => {
+    setOpen(true);
+  };
 
+  const handleClose = () => {
+    setOpen(false);
+  
+  };
+  function rand() {
+    return Math.round(Math.random() * 20) - 10;
+  }
+  
+  function getModalStyle() {
+    const top = 50 + rand();
+    const left = 50 + rand();
+  
+    return {
+      top: `${top}%`,
+      left: `${left}%`,
+      transform: `translate(-${top}%, -${left}%)`,
+    };
+  }
+  
+  const useStyles = makeStyles((theme: Theme) =>
+    createStyles({
+      paper: {
+        position: 'absolute',
+        width: 400,
+        backgroundColor: theme.palette.background.paper,
+        border: '2px solid #000',
+        boxShadow: theme.shadows[5],
+        padding: theme.spacing(2, 4, 3),
+      },
+    }),
+  );
+   const classes = useStyles();
+  const body = ( 
+    <div style={modalStyle} className={classes.paper}>
+      <h2 id="simple-modal-title">
+          { playlistName }
+        </h2>
+      {
+      <p id="simple-modal-description">
+          Playlist Tracks
+      </p>
+      }
+    </div>
+  );
 
   return (
     <Card>
@@ -36,7 +110,7 @@ const PlaylistCard = ({ playlistName }: any, { playlistId }: any) => {
           <Typography
             variant="body2"
             color="textSecondary"
-            onClick={handleClick}
+            onClick={handleOpen}
           >
             {playlistName}
           </Typography>
@@ -44,13 +118,18 @@ const PlaylistCard = ({ playlistName }: any, { playlistId }: any) => {
       </CardActionArea>
       <CardActions>
         <Button size="small" color="primary">
-          View Playlist
         </Button>
         <Button size="small" color="primary">
-          Delete Playlist
         </Button>
       </CardActions>
-      <PlaylistModal />
+      <Modal
+          open={open}
+          onClose={handleClose}
+          aria-labelledby="simple-modal-title"
+          aria-describedby="simple-modal-description"
+        >
+          {body}
+        </Modal>
     </Card>
   );
 };
