+++ client/src/redux/musicAppSlice.tsx
@@ -23,6 +23,7 @@ export interface MusicAppState {
     }>;
   }>;
   playlist: {
+    playlistId: string;
     username: string;
     playlistName: string;
     songs: Array<{
@@ -56,6 +57,7 @@ const initialState: MusicAppState = {
   album: { albumId: "", title: "", cover: "" },
   artist: { artistId: "", name: "", picture: "" },
   playlist: {
+    playlistId: "",
     username: "",
     playlistName: "",
     songs: [],
